Move exported submit function to new file
